--- RokuBack/config.js
@@ -1,6 +1,0 @@
-module.exports = {
-    host: 'localhost',
-    user: 'root',
-    password: '',
-    database: 'db_roku_flashback'
-}
