--- RokuBack/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "rokuback",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "start": "node app"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "express": "^4.17.3",
-    "mysql": "^2.18.1"
-  }
-}
