+++ config.js
@@ -1,0 +1,6 @@
+module.exports = {
+    host: 'localhost',
+    user: 'root',
+    password: '',
+    database: 'db_roku_flashback'
+}
